+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## May 28, 2020
+### New
+- error resolving by local message processing
+- `contracts.resolve.error` function for manual error resolving call
+- `contracts.process.transaction` function processing transaction to check errors and get output
+- `contracts.run.local` and `contracts.run.local` functions now have `fullRun` flag to emulate
+node transaction processing and calculate fees
+
 ## May 22, 2020
-### ton-client 0.23.1
 #### Fix
 - TON mnemonic functions didn't check validity of the seed phrase.